third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGPatternElement-dom-width-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'width' attribute of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.getAttribute('width') is "100"
PASS patternElement.getAttribute('width') is "50"
PASS successfullyParsed is true

TEST COMPLETE
